Format and lint issues
